backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
SQLAlchemy>=1.4,<2.0
psycopg2-binary>=2.9
python-dotenv>=0.21.0
passlib>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
stripe>=5.0.0
python-multipart>=0.0.6
pydantic>=1.10,<2.0